add back triton for potential rocm fallback. add import guards
